+++ PokerGameApp/PokerGameApp/Player.swift
@@ -1,6 +1,6 @@
 import Foundation
 
-class Player: Playable {
+class Player: PokerGame, Playable {
     var name: String
     private var cards = [Card]()
     var finalScore: [Card] {
@@ -9,13 +9,9 @@ class Player: Playable {
     
     private static var identifierFactory = 0
     
-    init() {
+    override init() {
         Player.identifierFactory += 1
         self.name = "참가자\(Player.identifierFactory)#"
-    }
-    
-    init(name: String) {
-        self.name = name
     }
     
     public func appendCard(_ card: Card) {
