+++ README.md
@@ -340,17 +340,17 @@ enum  BookFormat  {
 
 
 
-`registerPlayer(of:)` 에서 dealer도 해당시키기 위해 `player` 속성을 만들었습니다. 이후 step에서 다른 메소드에도 쓰일 수 있을 것 같습니다. 또한 dealer 자신을 player에 넣은 것은 최대한 비슷하게 구성해보고자 했습니다. Dealer의 역할은 게임플레이와 동시에 PokerGame 진행을 맡고 있기 때문에 Player에 들어간다는걸 일부러 코드로 써서 표시해서 혼동이 없게금 하고 싶었습니다. 하지만 코드 line이 늘어나는 부분은 아쉽게 생각합니다.
+리뷰주신 것을 보고 전체를 다시 보니 `players` 속성이 필요해보이는 것을 느꼈습니다. `registerPlayer(of:)` 에서 dealer도 해당시키기 위해 `player` 속성을 만들었습니다. 이후 step에서 다른 메소드에도 쓰일 수 있을 것 같습니다. 또한 dealer 자신을 player에 넣은 것은 최대한 비슷하게 구성해보고자 했습니다. Dealer의 역할은 게임플레이와 동시에 PokerGame 진행을 맡고 있기 때문에 Player에 들어간다는걸 일부러 코드로 써서 표시해서 혼동이 없게금 하고 싶었습니다. 하지만 코드 line이 늘어나는 부분은 아쉽게 생각합니다.
 
 
 
+#### Model 역할의 중요성
 
+> 출력하는 함수는 이렇게 딜러처럼 데이터를 다루는 타입에 넣지마세요.
+> 앱을 만들려고 하는 건데 이런 메소드는 앱에 화면을 표시하는데 사용하지 않는 함수가 됩니다.
+> 그래서 데이터를 전달하는 구조로 만드는게 좋습니다. 그래야 print가 아니라 다른 뷰에 표시도 가능해집니다.
 
-
-
-
-
-
+모델 역할의 큰 깨달음
